+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Updated comb_spec_searcher version for faster counting
 - Added a point placement strategy to partial row and col placement packs.
+- `TileScopePack.make_tracked` return a new pack with tracked fusion instead of
+  only adding the `AddAssumptionFactory` to the pack
 
 ## [2.4.1] - 2020-10-28
 ### Added